Throw an error if inline primitives refer to wrong primitive
Fixes issue #588. It should be noted that this merely "fixes" the issue
in the sense that users will be told to correct their mistake, while
they shouldn't have been able to make the mistake in the first place. We
should seriously think about ditching json-primitives entirely, and
offer a much more ergonomic API.

As a happy side-effect, this fixes the inline_primitives hack that was
around to support inline primitives. This commit also explicitly removes
the ability to use file references in inline primitives, as we can't
reliably determine where they are.