Fix misleading error message in "HasSingleDomain"
I haven't managed to reliably detect stuckness of constructs such as:

     System ~~ dom

In this case we would like to throw an ambiguous error, but we can't
because we're stuck. Strangely, similar strategies does seem to work for
HasSpecificDomain. Maybe it has something to do with the return types
of the different type families?